\name{berlineR-package}
\alias{berlineR-package}
\alias{berlineR}
\docType{package}
\title{
\packageTitle{berlineR}
}
\description{
\packageDescription{berlineR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{berlineR}
\packageIndices{berlineR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{berlineR}

Maintainer: \packageMaintainer{berlineR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
