% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f2n.R
\name{f2n}
\alias{f2n}
\title{factor to numeric}
\usage{
f2n(x)
}
\arguments{
\item{x}{an object of class factor}
}
\value{
the ORIGINAL numbering of your factors
}
\description{
If you ever have a factor where the levels do not
correspond with the defaults in R i.e (1,2)
this will convert back into what they originally are.
It is useful when you'd like to treat the factor as a vector
(as in using it with cbind, etc.)
}
\details{
This function converts it into a character string
and then into the numeric
}
\examples{
my_factor=factor(c(5,6,5,6),levels=c(5,6))
cbind(my_factor)
f2n(my_factor)
}

