% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trace.R
\name{Tr}
\alias{Tr}
\alias{trace}
\title{Trace of a matrix}
\usage{
Tr(x)
}
\arguments{
\item{x}{a square matrix}
}
\value{
the trace of the matrix \code{x}
}
\description{
Trace of a matrix
}
\details{
Can you believe that this is not available as a function in base?
  This function computes the trace of a matrix by first getting the diagonal
  elements, and then summming.
}
\examples{
Tr(diag(5))
}

